Remove unreachable code
I added this earlier when I started doing /len(snapshots) and I was
concerned about a DivByZero exception when someone had just made a new
dataset.

BUT, this can't happen, because zfs list -t snapshot WON'T EVEN
MENTION datasets with no snapshots.

Likewise, datasets with only snapshots-that-we-ignore won't result in
the creation of an empty list, because of how defaultdict() works.