expire: destroy (not list) recursively
This basically makes expire match snapshot, operating recursively from the root specified by --dataset.

This should be substantially simpler/faster because it makes recursion zfs(8)'s problem.

This should make it harder for individual datasets to get out-of-sync with each other.

HOWEVER, it also means that if someone manually creates a snapshot of a single dataset, it will NEVER BE EXPIRED.
For example,

    cyber-backup-zfs --action=snapshot --pool=A/A       # creates A/A/home@1
    cyber-backup-zfs --action=snapshot --pool=A/A/home  # creates A/A/home@2
    cyber-backup-zfs --action=expire   --pool=A/A       # will no longer see the @2!

I think this is an overall win in terms of reliability AND sanity.